streamlit-app/requirements.txt
streamlit>=1.30.0
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-chroma>=0.2.0
langchain-community>=0.3.0
chromadb>=0.4.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
